refactor(lowering): extract error type and helpers into separate modules
