+++ crates/js-lowering/src/lower/mod.rs
@@ -1,33 +1,17 @@
 mod control_flow;
 mod declarations;
+mod error;
 mod expressions;
-
-use std::fmt;
+mod helpers;
 
 use ehrmantraut_core::ir::{
-  Block, DeclKind, ExpressionStatement, IrExpr, IrModule, IrNode, OpaqueExpr, OpaqueNode,
-  ScopeLevel,
+  Block, ExpressionStatement, IrExpr, IrModule, IrNode, OpaqueExpr, OpaqueNode,
 };
 
 use crate::cst::CstNode;
 use crate::parser::Language;
 
-#[derive(Debug)]
-pub enum LowerError {
-  UnexpectedNode { kind: String, expected: String },
-}
-
-impl fmt::Display for LowerError {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    match self {
-      LowerError::UnexpectedNode { kind, expected } => {
-        write!(f, "unexpected node '{kind}', expected {expected}")
-      }
-    }
-  }
-}
-
-impl std::error::Error for LowerError {}
+pub use error::LowerError;
 
 pub fn lower(cst: &CstNode, language: Language, source: &str) -> Result<IrModule, LowerError> {
   let mut lowerer = JsLowerer::new(language, source);
@@ -233,83 +217,5 @@ impl JsLowerer {
       span: node.span,
       body,
     })
-  }
-
-  // ── Helpers ──────────────────────────────────────────────────────
-
-  pub(crate) fn first_named_child<'a>(&self, node: &'a CstNode) -> Option<&'a CstNode> {
-    node.children.iter().find(|c| c.named)
-  }
-
-  pub(crate) fn child_by_field<'a>(&self, node: &'a CstNode, field: &str) -> Option<&'a CstNode> {
-    node
-      .children
-      .iter()
-      .find(|c| c.field_name.as_deref() == Some(field))
-  }
-
-  pub(crate) fn node_text(&self, node: &CstNode) -> String {
-    if let Some(text) = &node.text {
-      return text.clone();
-    }
-    // For branch nodes, reconstruct from source if available
-    if !self.source.is_empty() {
-      let start = node.span.start.offset as usize;
-      let end = node.span.end.offset as usize;
-      if start <= end && end <= self.source.len() {
-        return self.source[start..end].to_string();
-      }
-    }
-    // Fallback: concatenate leaf text
-    self.collect_text(node)
-  }
-
-  fn collect_text(&self, node: &CstNode) -> String {
-    if let Some(text) = &node.text {
-      return text.clone();
-    }
-    node
-      .children
-      .iter()
-      .map(|c| self.collect_text(c))
-      .collect::<Vec<_>>()
-      .join("")
-  }
-
-  pub(crate) fn strip_quotes(text: String) -> String {
-    if text.len() >= 2
-      && ((text.starts_with('"') && text.ends_with('"'))
-        || (text.starts_with('\'') && text.ends_with('\''))
-        || (text.starts_with('`') && text.ends_with('`')))
-    {
-      text[1..text.len() - 1].to_string()
-    } else {
-      text
-    }
-  }
-
-  pub(crate) fn decl_kind_and_scope(kind_str: &str) -> (Option<DeclKind>, Option<ScopeLevel>) {
-    match kind_str {
-      "let" => (Some(DeclKind::Let), Some(ScopeLevel::Block)),
-      "const" => (Some(DeclKind::Const), Some(ScopeLevel::Block)),
-      "var" => (Some(DeclKind::Var), Some(ScopeLevel::Function)),
-      _ => (None, None),
-    }
-  }
-
-  fn is_ts_type_node(kind: &str) -> bool {
-    matches!(
-      kind,
-      "type_annotation"
-        | "type_parameters"
-        | "type_arguments"
-        | "interface_declaration"
-        | "type_alias_declaration"
-        | "enum_declaration"
-        | "as_expression"
-        | "satisfies_expression"
-        | "non_null_expression"
-        | "type_assertion"
-    )
   }
 }
